13. Roman to Integer
